### 2020-01-17 #2021 @kyleniemeyer
Hi @embruna, thanks for your submission to JOSS. To confirm, this has already been reviewed by rOpenSci? If so, can you share the review thread here?
