### 2020-01-17 #2021 @embruna
This package has been reviewed by rOpenSci: [https://github.com/ropensci/software-review/issues/256](https://github.com/ropensci/software-review/issues/256), @maelle
