### 2020-01-17 #2021 @embruna
> One issue: your repository is missing a LICENSE file, corresponding to a valid open-source software license

Thanks @kyleniemeyer.  The Description notes it's licensed under GPL-3, but I'll add a corresponding file right away.
