### 2020-01-17 9f501b9
Update DESCRIPTION
added note to "License" entry referring to new LICENSE file
