+++ DESCRIPTION
@@ -33,7 +33,7 @@ Description: Tools to parse and organize reference records downloaded from the
     the names of authors, geocode their locations, and generate/visualize 
     coauthorship networks. This package has been peer-reviewed by 
     rOpenSci (v. 1.0).
-License: GPL-3
+License: GPL-3 + file LICENSE
 URL: https://github.com/ropensci/refsplitr, https://docs.ropensci.org/refsplitr/
 BugReports: https://github.com/ropensci/refsplitr/issues
 Depends: 
