### 2020-01-17 1c4407f
Update README.Rmd
added GPL3 Badge
