+++ README.Rmd
@@ -5,7 +5,11 @@ title: refsplitr
 
 <img src="man/figures/refsplitrhex.png" height="200" align="right">
 
-[![Project Status: Active The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)    [![Build Status](https://travis-ci.org/ropensci/refsplitr.svg?branch=master)](https://travis-ci.org/ropensci/refsplitr)    [![](https://badges.ropensci.org/256_status.svg)](https://github.com/ropensci/onboarding/issues/256)   [![Coverage Status](https://coveralls.io/repos/github/ropensci/refsplitr/badge.svg?branch=master)](https://coveralls.io/github/ropensci/refsplitr?branch=master)
+[![Project Status: Active The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
+[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+[![Build Status](https://travis-ci.org/ropensci/refsplitr.svg?branch=master)](https://travis-ci.org/ropensci/refsplitr)
+[![](https://badges.ropensci.org/256_status.svg)](https://github.com/ropensci/onboarding/issues/256)   
+[![Coverage Status](https://coveralls.io/repos/github/ropensci/refsplitr/badge.svg?branch=master)](https://coveralls.io/github/ropensci/refsplitr?branch=master)
 
 refsplitr: author name disambiguation, author georeferencing, and mapping of coauthorship networks with Web of Science data. 
 
