### 2020-01-17 5375699
Update README.md
added a badge for GPL3
