### 2020-01-17 7fc5cb3
Update README.Rmd
corrected spacing of badges
