### 2020-01-17 #2021 @embruna
@kyleniemeyer It's now there, along with an edit in the DESCRIPTION pointing to it (`License: GPL-3 + file LICENSE`).
