### 2020-01-17 #2028 @whedon
[REVIEW]: refsplitr: author name disambiguation, author georeferencing, and mapping of coauthorship networks with 'Web of Science' data.
**Submitting author:** @embruna (<a href="http://orcid.org/0000-0003-3381-8477">Emilio M. Bruna</a>)
**Repository:** <a href="https://github.com/ropensci/refsplitr" target ="_blank">https://github.com/ropensci/refsplitr</a>
**Version:** v1.0.0
