### 2020-01-21 #2028 @kyleniemeyer
This was reviewed by rOpenSci: ropensci/software-review#256, @maelle
