### 2020-01-22 #2028 @embruna
@kyleniemeyer Thanks! I froze the version submitted to JOSS at Zenodo last week ([doi: 10.5281/zenodo.3608285](https://zenodo.org/record/3608285#.XihPahdKg6c), and was going freeze the 1.0 version upon acceptance so that I could include the link and citation of the JOSS paper. Is that ok?
