### 2020-01-22 #2028 @arfon
@embruna - one quick question on this submission. Is there a reason why you have `formatted_doi:
` as a field in your YAML header? Having this as a blank entry caused some issues in our publishing pipeline earlier.
