### 2020-01-22 #2028 @kyleniemeyer
@arfon I wondered why the DOI didn't resolve earlier...
