### 2020-01-22 #2028 @arfon
> @arfon I wondered why the DOI didn't resolve earlier...

😄 yeah. `formatted_doi` is one of the variables we pass into the Pandoc template. Having this in the YAML header of the `paper.md` file overrides the values that Whedon was passing in.
