### 2020-01-23 #2028 @embruna
Yup, it was in the template...but not anymore.  Deleted.
