### 2020-01-24 #2028 @embruna
That would be useful - either removal from the template OR a note that it should be deleted.
