### 2020-01-24 #2028 @arfon
> That would be useful - either removal from the template OR a note that it should be deleted.

So this field _is_ included in a template somewhere? If so, could you please point me to it?
